refactor:  move model
